+++ app/cmd/site/main.go
@@ -1,0 +1,7 @@
+package main
+
+import "github.com/desertthunder/dotfiler/internal/cli"
+
+func main() {
+	cli.Site.Execute()
+}